+++ tests/doc/capi/chfl_cell/copy.c
@@ -6,7 +6,7 @@
 
 int main() {
     // [example]
-    CHFL_CELL* cell = chfl_cell((chfl_vector3d){10, 10, 10});
+    CHFL_CELL* cell = chfl_cell((chfl_vector3d){10, 10, 10}, NULL);
     CHFL_CELL* copy = chfl_cell_copy(cell);
 
     if (copy == NULL) {