+++ tests/doc/capi/chfl_cell/from_matrix.c
@@ -7,15 +7,22 @@
 
 int main() {
     // [example]
-    CHFL_CELL* cell = chfl_cell((chfl_vector3d){10, 11, 12}, NULL);
+    chfl_vector3d matrix[3] = {
+        {10, 0, 0},
+        {0, 12, 0},
+        {0, 0, 15}
+    };
 
-    chfl_vector3d lengths = {0, 0, 0};
+    CHFL_CELL* cell = chfl_cell_from_matrix(matrix);
+
+    chfl_vector3d lengths;
     chfl_cell_lengths(cell, lengths);
     assert(lengths[0] == 10);
-    assert(lengths[1] == 11);
-    assert(lengths[2] == 12);
+    assert(lengths[1] == 12);
+    assert(lengths[2] == 15);
 
     chfl_free(cell);
+
     // [example]
     return 0;
 }