+++ tests/doc/capi/chfl_cell/shape.c
@@ -7,7 +7,7 @@
 
 int main() {
     // [example]
-    CHFL_CELL* cell = chfl_cell((chfl_vector3d){10, 10, 10});
+    CHFL_CELL* cell = chfl_cell((chfl_vector3d){10, 10, 10}, NULL);
 
     chfl_cellshape shape;
     chfl_cell_shape(cell, &shape);